TBR-0006: Update Docker configuration (#116)

+++ scripts/build.sh
@@ -1,8 +1,13 @@
 #!/usr/bin/env bash
 set -e
 
-echo "==> Building Docker dev image (if needed)..."
-docker compose build dev
+if [ "$1" == "prod" ]; then
+    echo "==> Building Docker prod image..."
+    docker compose build prod
+else
+    echo "==> Building Docker dev image (if needed)..."
+    docker compose build dev
 
-echo "==> Compiling backend code using Clang + Ninja..."
-docker compose run --rm dev sh -c "CC=clang CXX=clang++ cmake -B build -S . -G Ninja && cmake --build build -j$(nproc)"
+    echo "==> Compiling backend code using Clang + Ninja..."
+    docker compose run --rm dev sh -c "CC=clang CXX=clang++ cmake -B build -S . -G Ninja && cmake --build build -j$(nproc)"
+fi
